remove wrapped defined types for dependencies
Wrap types as defined in dependencies as those API exposed to users seem
to be redundancy as we need to do double work. Users can directly use
OpenGL APIs thus it's better for users to still use types that are already
known to them.

In anothel thought, if we will actually do unified wrapped of defined types as defined
in dependencies, we probably should went far to wrap those APIs as well and expose
only certain functionality.

But as for current state, we take in library to offer a certain functionality
immediately and the goal is not to build a complete engine yet. Thus this
makes sense.