Refactor inclusion of External.h header file and minor refactor
Use similar approach as done in stb_image by using #define ...
to indicate which part of source code to be included.

As aim to put most (if not all) of the external dependencies' header
files inclusion into External.h which will in turn included by lgl code
based (possibly across the places which need to use). Then this will
reduce the time used in compilation to include only things needed.